requirements.txt
# Backend dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
google-generativeai>=0.3.0
python-dotenv>=1.0.0
radon>=6.0.0

# Frontend dependencies
streamlit>=1.28.0
requests>=2.31.0